Ajout de la nouvelle commande EasyUse OK
