+++ src/main/java/ch/heigvd/labo/Main.java
@@ -1,19 +1,16 @@
 package ch.heigvd.labo;
 
 import java.util.concurrent.Callable;
+
+import ch.heigvd.labo.command.*;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
-import ch.heigvd.labo.command.Build;
-import ch.heigvd.labo.command.Clean;
-import ch.heigvd.labo.command.New;
-import ch.heigvd.labo.command.Serve;
-import ch.heigvd.labo.command.Init;
 
 @Command(
         name="Main",
         description = "Programme DGPW - Website Generator" ,
-        subcommands = {New.class, Clean.class, Build.class, Serve.class, Init.class} )
+        subcommands = {New.class, Clean.class, Build.class, Serve.class, Init.class, EasyUse.class })
 
 public class Main implements Callable<Integer>{
     @Option(names = { "--version", "-v" }, description = "Display the version and exit")
