+++ src/main/java/ch/heigvd/labo/command/Serve.java
@@ -22,41 +22,6 @@ public class Serve implements Callable {
     private File dirStatic;
 
     @Override public Integer call(){
-        int result = 0;
-        if(dirStatic != null) {
-            // Tous les autres tests sont effectués dans les classes des commandes directement
-            // Nous appelons uniquement la commande en testant la valeur de retour de la commande.
-            if (initOption != null && initOption) {
-                result = new CommandLine(new Init()).execute("-d", dirStatic.getPath());
-                if(result != 0) {
-                    System.out.println("Une erreur s'est produite durant le init");
-                    return 1;
-                }
-            }
-
-            else if (newOption != null && newOption) {
-                if(filePage != null) {
-                    result = new CommandLine(new New()).execute("-d", dirStatic.getPath(), "-f", filePage.getName());
-                    if (result != 0) {
-                        System.out.println("Une erreur s'est produite durant le new");
-                        return 2;
-                    }
-                }
-                else {
-                    System.out.println("Merci de renseigner le nom du répertoire et/ou du fichier à créer : \n-d [nom du répertoire] \n-f [nom du fichier]");
-                    return 3;
-                }
-            }
-
-            result = new CommandLine(new Build()).execute("-d", dirStatic.getPath());
-            if(result != 0) {
-                System.out.println("Une erreur s'est produite durant le build");
-                return 4;
-            }
-            System.out.println("Vous pouvez accéder au site web via " + Utility.LOCALHOST+Utility.DIR_ROOT+dirStatic.getPath()+Utility.DIR_BUILD +"/index.html");
-            return 0;
-        }
-        System.out.println("Merci de renseigner le nom du répertoire et/ou du fichier à créer : \n-d [nom du répertoire] \n-f [nom du fichier]");
-        return 3;
+        return 1;
     }
 }
